Add support for custom URL converters
This change adds support for custom URL converters so long as they
return an Enum. It also includes a bit of an overhaul of the path
parsing code as it had a few mistakes.

New feature:
If a custom URL converter has a type annotation on it's `to_python`
method and that method returns an Enum then generate an appropriate
schema for it. Otherwise fall back to a default string schema. This
results in there always being a schema otherwise the generated OpenAPI
spec is not valid.

Fixes:
The "any" converter [1] does not expect an array of enums. It expects
*one* of the given items.

The "string" converter [2] expects the argument keys to be lowercase.

Having a null schema is not valid according to the OpenAPI spec. This
change falls back to a default string schema as ultimately the path is a
string.

[1] https://werkzeug.palletsprojects.com/en/2.3.x/routing/#werkzeug.routing.AnyConverter
[2] https://werkzeug.palletsprojects.com/en/2.3.x/routing/#werkzeug.routing.UnicodeConverter