Support stacked routes
A route function can be decorated more than once in flask. Currently
this results in two OpenAPI operations with the same ID which is not
valid. This change uses the route's `endpoint` which defaults to the
name of the wrapped function.

See: https://flask.palletsprojects.com/en/stable/api/#flask.Flask.add_url_rule