Allow operationId to be customized
This effectively reverts #87 as it had unintended consequences
when used with blueprints. The endpoint for a blueprint by default
is `app_name.blueprint_name.function_name`. This means that #83
would result in significant changes to existing OpenAPI specs.
This change instead keeps the operationId the same by default but
offers a route to overriding it.